Remove any Unbind method and introduce none objects, remove deprecates.
This commit removes any Unbind method replacing it with a Bind of
a none object. None objects are just 0 constants of the respective
OpenGL object (texture, VBO and so on). Unbind methods are dangerous
and counter-intuitive, an Unbind of an unbound object causes the
actually bound object to be unbound, which is dangerous and undesired.
None object obviate the issue by using the actual OpenGL approach.
This commit also removes previously deprecated functionality to
do the same thing.